Move avatar and user menu to subcomponent; integrate Google Login Btn
