+++ frontend/src/components/LoginAvatarMenu.jsx
@@ -1,0 +1,66 @@
+import React, { useContext } from "react";
+import { UserContext } from "../contexts/UserContext";
+import {
+  Box,
+  Tooltip,
+  IconButton,
+  Avatar,
+  Menu,
+  MenuItem,
+  Typography,
+} from "@mui/material";
+
+const settings = ["Profile", "Dashboard", "Logout"];
+
+export default function LoginAvatarMenu() {
+  const user = useContext(UserContext);
+
+  const [anchorElUser, setAnchorElUser] = React.useState(null);
+
+  const handleOpenUserMenu = (event) => {
+    setAnchorElUser(event.currentTarget);
+  };
+
+  const handleCloseUserMenu = () => {
+    setAnchorElUser(null);
+  };
+  if (user) {
+    return (
+      <Box sx={{ flexGrow: 0 }}>
+        <Tooltip title="Open settings">
+          <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
+            <Avatar alt={user.name} src={user.picture} />
+          </IconButton>
+        </Tooltip>
+        <Menu
+          sx={{ mt: "45px" }}
+          id="menu-appbar"
+          anchorEl={anchorElUser}
+          anchorOrigin={{
+            vertical: "top",
+            horizontal: "right",
+          }}
+          keepMounted
+          transformOrigin={{
+            vertical: "top",
+            horizontal: "right",
+          }}
+          open={Boolean(anchorElUser)}
+          onClose={handleCloseUserMenu}
+        >
+          {settings.map((setting) => (
+            <MenuItem key={setting} onClick={handleCloseUserMenu}>
+              <Typography textAlign="center">{setting}</Typography>
+            </MenuItem>
+          ))}
+        </Menu>
+      </Box>
+    );
+  } else {
+    return (
+      <Box sx={{ flexGrow: 0 }}>
+        <div id="googleLoginBtn"></div>
+      </Box>
+    );
+  }
+}
