+++ frontend/src/components/NavBar.jsx
@@ -1,33 +1,10 @@
 import * as React from "react";
-import {
-  AppBar,
-  Box,
-  Toolbar,
-  IconButton,
-  Typography,
-  Menu,
-  Container,
-  Avatar,
-  Button,
-  Tooltip,
-  MenuItem
-} from "@mui/material";
-import { MenuRounded } from "@mui/icons-material"
+import { AppBar, Box, Toolbar, Container, Button } from "@mui/material";
+import LoginAvatarMenu from "./LoginAvatarMenu";
 
 const pages = ["Experiences", "Page 1", "Page 2"];
-const settings = ["Profile", "Dashboard", "Logout"];
 
 function ResponsiveAppBar() {
-  const [anchorElUser, setAnchorElUser] = React.useState(null);
-
-  const handleOpenUserMenu = (event) => {
-    setAnchorElUser(event.currentTarget);
-  };
-
-  const handleCloseUserMenu = () => {
-    setAnchorElUser(null);
-  };
-
   const handleLinkClick = (event) => {
     console.log(event.target.id);
   };
@@ -36,7 +13,7 @@ function ResponsiveAppBar() {
     <AppBar position="static">
       <Container maxWidth="xl">
         <Toolbar disableGutters>
-          <img src={"./logo.png"} width={80} height={56}/>
+          <img src={"./logo.png"} width={80} height={56} />
 
           <Box sx={{ flexGrow: 1, display: { xs: "none", md: "flex" } }}>
             {pages.map((page, i) => (
@@ -46,40 +23,11 @@ function ResponsiveAppBar() {
                 onClick={handleLinkClick}
                 sx={{ my: 2, color: "white", display: "block" }}
               >
-                {page} 
+                {page}
               </Button>
             ))}
           </Box>
-
-          <Box sx={{ flexGrow: 0 }}>
-            <Tooltip title="Open settings">
-              <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-                <Avatar alt="Remy Sharp" src="/static/images/avatar/2.jpg" />
-              </IconButton>
-            </Tooltip>
-            <Menu
-              sx={{ mt: "45px" }}
-              id="menu-appbar"
-              anchorEl={anchorElUser}
-              anchorOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              open={Boolean(anchorElUser)}
-              onClose={handleCloseUserMenu}
-            >
-              {settings.map((setting) => (
-                <MenuItem key={setting} onClick={handleCloseUserMenu}>
-                  <Typography textAlign="center">{setting}</Typography>
-                </MenuItem>
-              ))}
-            </Menu>
-          </Box>
+          <LoginAvatarMenu />
         </Toolbar>
       </Container>
     </AppBar>
